model_training/requirements.txt
gym >= 0.26
numpy==1.24.3
jax==0.4.13
distrax==0.1.2
flax==0.7.0
ml_collections >= 0.1.0
tqdm >= 4.60.0
chex==0.1.6
optax==0.1.5
absl-py >= 0.12.0
scipy >= 1.6.0
wandb >= 0.12.14
tensorflow==2.14.0
einops >= 0.6.1
imageio >= 2.31.1
moviepy >= 1.0.3
pre-commit == 3.3.3
overrides==7.7.0
tensorflow_probability==0.22.1
